+++ werk-cli/main.rs
@@ -263,30 +263,6 @@ async fn try_main(args: Args) -> Result<(), Error> {
         return Ok(());
     }
 
-    let target = args
-        .target
-        .clone()
-        .or_else(|| workspace.default_target.clone());
-    let Some(target) = target else {
-        return Err(Error::NoTarget);
-    };
-
-    let runner = Runner::new(&workspace);
-    let result = runner.build_or_run(&target).await;
-
-    let write_cache = match result {
-        Ok(_) => true,
-        Err(ref err) => err.error.should_still_write_werk_cache(),
-    };
-
-    if write_cache {
-        if let Err(err) = workspace.finalize().await {
-            eprintln!("Error writing `.werk-cache`: {err}")
-        }
-    }
-
-    std::mem::drop(runner);
-
     if args.watch {
         autowatch_loop(
             std::time::Duration::from_millis(args.watch_delay),
@@ -299,14 +275,25 @@ async fn try_main(args: Args) -> Result<(), Error> {
         .await?;
         Ok(())
     } else {
-        result.map(|_| ()).map_err(print_error)
+        let target = args
+            .target
+            .clone()
+            .or_else(|| workspace.default_target.clone());
+        let Some(target) = target else {
+            workspace
+                .render
+                .runner_message("No configured default target");
+            return Err(Error::NoTarget);
+        };
+
+        run(&workspace, &target).await.map(|_| ())
     }
 }
 
 async fn autowatch_loop(
     timeout: std::time::Duration,
     // The initial workspace built by main(). Must be finalize()d.
-    workspace: Workspace,
+    mut workspace: Workspace,
     werkfile: Absolute<std::path::PathBuf>,
     // Target to keep building
     target_from_args: Option<String>,
@@ -332,11 +319,30 @@ async fn autowatch_loop(
         }
     }));
     let workspace_dir = workspace.project_root().to_path_buf();
-    std::mem::drop(workspace);
 
     let mut settings = settings.clone();
 
+    let mut target = target_from_args
+        .clone()
+        .or_else(|| workspace.default_target.clone());
+
     loop {
+        if target.is_none() {
+            render.runner_message("No configured default target");
+            watch_set = watch_manifest.clone();
+        }
+
+        // Start the notifier.
+        let notifier = make_notifier_for_files(&watch_set, notification_sender.clone(), timeout)?;
+        let notification_recv = notification_receiver.recv();
+
+        // Build or rebuild the target!
+        let runner = if let Some(target) = target.as_ref() {
+            run(&workspace, target).await.map(Some)
+        } else {
+            Ok(None)
+        };
+
         if watch_set == watch_manifest {
             render.runner_message("Watching manifest for changes, press Ctrl-C to stop");
         } else {
@@ -346,9 +352,6 @@ async fn autowatch_loop(
             ));
         }
 
-        // Start the notifier.
-        let notifier = make_notifier_for_files(&watch_set, notification_sender.clone(), timeout)?;
-        let notification_recv = notification_receiver.recv();
         let ctrlc_recv = ctrlc_receiver.recv();
         smol::pin!(notification_recv);
         smol::pin!(ctrlc_recv);
@@ -362,6 +365,8 @@ async fn autowatch_loop(
                 }
             }
         }
+
+        std::mem::drop(runner);
 
         // Stop the notifier again immediately. TODO: Consider if it makes sense to reuse it.
         notifier.stop();
@@ -430,7 +435,7 @@ async fn autowatch_loop(
             settings.output_directory = out_dir;
         }
 
-        let mut workspace =
+        workspace =
             match Workspace::new(io.clone(), render.clone(), workspace_dir.clone(), &settings) {
                 Ok(workspace) => workspace,
                 Err(err) => {
@@ -453,14 +458,9 @@ async fn autowatch_loop(
             }
         }
 
-        let target = target_from_args
+        target = target_from_args
             .clone()
             .or_else(|| workspace.default_target.clone());
-        let Some(target) = target else {
-            render.runner_message("No configured default target");
-            watch_set = watch_manifest.clone();
-            continue;
-        };
 
         // Update the watchset.
         watch_set.clear();
@@ -472,25 +472,25 @@ async fn autowatch_loop(
                 None
             }
         }));
+    }
+}
 
-        // Finally, rebuild the target!
-        let runner = Runner::new(&workspace);
-        let write_cache = match runner.build_or_run(&target).await {
-            Ok(_) => true,
-            Err(err) => {
-                let write_cache = err.error.should_still_write_werk_cache();
-                print_error(err);
-                write_cache
-            }
-        };
+async fn run<'w>(workspace: &'w Workspace, target: &str) -> Result<Runner<'w>, Error> {
+    let runner = Runner::new(workspace);
+    let result = runner.build_or_run(target).await;
 
-        if write_cache {
-            if let Err(err) = workspace.finalize().await {
-                eprintln!("Error writing `.werk-cache`: {err}");
-                return Err(err.into());
-            }
+    let write_cache = match result {
+        Ok(_) => true,
+        Err(ref err) => err.error.should_still_write_werk_cache(),
+    };
+
+    if write_cache {
+        if let Err(err) = workspace.finalize().await {
+            eprintln!("Error writing `.werk-cache`: {err}")
         }
     }
+
+    result.map(|_| runner).map_err(print_error)
 }
 
 fn make_notifier_for_files(
