+++ werk-runner/runner.rs
@@ -15,6 +15,14 @@ use crate::{
     ir,
 };
 
+mod command;
+mod dep_chain;
+mod task;
+
+pub use command::*;
+pub use dep_chain::*;
+pub use task::*;
+
 /// Workspace-wide runner state.
 pub(crate) struct RunnerState {
     concurrency_limit: smol::lock::Semaphore,
@@ -81,125 +89,6 @@ impl BuildStatus {
                 }
             }
             BuildStatus::Ignore(_) => None,
-        }
-    }
-}
-
-enum TaskStatus {
-    Built(Result<BuildStatus, Error>),
-    Pending(Vec<oneshot::Sender<Result<BuildStatus, Error>>>),
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum TaskId {
-    Task(Symbol),
-    // TODO: When recipes can build multiple files, this needs to change to some
-    // ID that encapsulates the "recipe instance" rather than the path of a
-    // single target.
-    Build(Absolute<SymPath>),
-}
-
-impl TaskId {
-    pub fn command(s: impl Into<Symbol>) -> Self {
-        let name = s.into();
-        debug_assert!(!name.as_str().starts_with('/'));
-        TaskId::Task(name)
-    }
-
-    pub fn build(p: impl AsRef<Absolute<werk_fs::Path>>) -> Self {
-        TaskId::Build(Absolute::symbolicate(p))
-    }
-
-    pub fn try_build<P>(p: P) -> Result<Self, P::Error>
-    where
-        P: TryInto<Absolute<werk_fs::PathBuf>>,
-    {
-        let path = p.try_into()?;
-        Ok(TaskId::build(path))
-    }
-
-    #[inline]
-    #[must_use]
-    pub fn is_command(&self) -> bool {
-        matches!(self, TaskId::Task(_))
-    }
-
-    #[inline]
-    #[must_use]
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            TaskId::Task(task) => task.as_str(),
-            TaskId::Build(build) => build.as_inner().as_str(),
-        }
-    }
-
-    #[inline]
-    #[must_use]
-    pub fn as_path(&self) -> Option<&Absolute<werk_fs::Path>> {
-        if let TaskId::Build(build) = self {
-            Some(build.as_path())
-        } else {
-            None
-        }
-    }
-
-    #[inline]
-    #[must_use]
-    pub fn short_name(&self) -> &'static str {
-        match self {
-            TaskId::Task(task) => task.as_str(),
-            TaskId::Build(path) => {
-                let Some((_prefix, filename)) = path
-                    .as_inner()
-                    .as_str()
-                    .rsplit_once(werk_fs::Path::SEPARATOR)
-                else {
-                    // The path is absolute.
-                    unreachable!()
-                };
-                filename
-            }
-        }
-    }
-}
-
-impl std::fmt::Display for TaskId {
-    #[inline]
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(self.as_str())
-    }
-}
-
-enum TaskSpec<'a> {
-    Recipe(ir::RecipeMatch<'a>),
-    CheckExists(Absolute<werk_fs::PathBuf>),
-    /// Check if the file exists, but don't emit an error if it doesn't. This
-    /// applies to dependencies discovered through depfiles, where the depfile
-    /// may be outdated (from a previous build).
-    ///
-    /// If the file does not exist, the task will be considered outdated.
-    CheckExistsRelaxed(Absolute<werk_fs::PathBuf>),
-}
-
-enum DepfileSpec<'a> {
-    /// The depfile is explicitly generated by a recipe.
-    Recipe(ir::BuildRecipeMatch<'a>),
-    /// The depfile is implicitly generated by the associated command recipe.
-    ImplicitlyGenerated(Absolute<werk_fs::PathBuf>),
-}
-
-impl TaskSpec<'_> {
-    pub fn to_task_id(&self) -> TaskId {
-        match self {
-            TaskSpec::Recipe(ir::RecipeMatch::Build(build_recipe_match)) => {
-                TaskId::build(build_recipe_match.target_file.clone())
-            }
-            TaskSpec::Recipe(ir::RecipeMatch::Task(task_recipe_match)) => {
-                TaskId::command(task_recipe_match.name)
-            }
-            TaskSpec::CheckExists(path_buf) | TaskSpec::CheckExistsRelaxed(path_buf) => {
-                TaskId::build(path_buf.clone().into_boxed_path())
-            }
         }
     }
 }
@@ -980,163 +869,5 @@ impl<'a> Inner<'a> {
             TaskSpec::CheckExists(path) => self.check_exists(&path),
             TaskSpec::CheckExistsRelaxed(path) => Ok(self.check_exists_relaxed(&path)),
         }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum RunCommand {
-    Shell(ShellCommandLine),
-    Write(Absolute<std::path::PathBuf>, Vec<u8>),
-    // We don't know yet if the source file is in the workspace or output
-    // directory, so we will resolve the path when running it.
-    Copy(Absolute<werk_fs::PathBuf>, Absolute<std::path::PathBuf>),
-    Info(DiagnosticSpan, String),
-    Warn(DiagnosticSpan, String),
-    // Path is always in the output directory. They don't need to exist.
-    Delete(DiagnosticSpan, Vec<Absolute<std::path::PathBuf>>),
-    Touch(DiagnosticSpan, Vec<Absolute<std::path::PathBuf>>),
-    SetCapture(bool),
-    SetEnv(String, String),
-    RemoveEnv(String),
-}
-
-impl std::fmt::Display for RunCommand {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            RunCommand::Shell(shell_command_line) => shell_command_line.fmt(f),
-            RunCommand::Write(path_buf, vec) => {
-                write!(f, "write {} ({} bytes)", path_buf.display(), vec.len())
-            }
-            RunCommand::Copy(from, to) => {
-                write!(f, "copy '{}' to '{}'", from, to.display())
-            }
-            RunCommand::Info(_, message) => {
-                write!(f, "info \"{}\"", message.escape_default())
-            }
-            RunCommand::Warn(_, message) => {
-                write!(f, "warn \"{}\"", message.escape_default())
-            }
-            RunCommand::Delete(_, paths) => {
-                write!(f, "delete ")?;
-                if paths.len() == 1 {
-                    write!(f, "{}", paths[0].display())
-                } else {
-                    write!(f, "[")?;
-                    for (i, p) in paths.iter().enumerate() {
-                        if i != 0 {
-                            write!(f, ", ")?;
-                        }
-                        write!(f, "{}", p.display())?;
-                    }
-                    write!(f, "]")
-                }
-            }
-            RunCommand::Touch(_, paths) => {
-                write!(f, "touch ")?;
-                if paths.len() == 1 {
-                    write!(f, "{}", paths[0].display())
-                } else {
-                    write!(f, "[")?;
-                    for (i, p) in paths.iter().enumerate() {
-                        if i != 0 {
-                            write!(f, ", ")?;
-                        }
-                        write!(f, "{}", p.display())?;
-                    }
-                    write!(f, "]")
-                }
-            }
-            RunCommand::SetCapture(value) => write!(f, "set_capture = {value}"),
-            RunCommand::SetEnv(key, value) => write!(f, "env {key} = {value}"),
-            RunCommand::RemoveEnv(key) => write!(f, "env-remove {key}"),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-struct DepChainEntry<'a> {
-    parent: DepChain<'a>,
-    this: TaskId,
-}
-
-#[derive(Debug, Clone, Copy)]
-enum DepChain<'a> {
-    Empty,
-    Owned(&'a OwnedDependencyChain),
-    Ref(&'a DepChainEntry<'a>),
-}
-
-impl<'a> DepChain<'a> {
-    fn collect_vec(&self) -> Vec<TaskId> {
-        match self {
-            DepChain::Empty => Vec::new(),
-            DepChain::Owned(owned) => owned.vec.clone(),
-            DepChain::Ref(parent) => parent.collect_vec(),
-        }
-    }
-
-    pub fn contains(&self, task: TaskId) -> bool {
-        match self {
-            DepChain::Empty => false,
-            DepChain::Owned(owned) => owned.vec.contains(&task),
-            DepChain::Ref(parent) => parent.contains(task),
-        }
-    }
-
-    pub fn push<'b>(self, task: TaskId) -> DepChainEntry<'b>
-    where
-        'a: 'b,
-    {
-        DepChainEntry {
-            parent: self,
-            this: task,
-        }
-    }
-}
-
-impl DepChainEntry<'_> {
-    fn collect(&self) -> OwnedDependencyChain {
-        OwnedDependencyChain {
-            vec: self.collect_vec(),
-        }
-    }
-
-    fn collect_vec(&self) -> Vec<TaskId> {
-        let mut vec = self.parent.collect_vec();
-        vec.push(self.this);
-        vec
-    }
-
-    fn contains(&self, task_id: TaskId) -> bool {
-        if self.this == task_id {
-            true
-        } else {
-            self.parent.contains(task_id)
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct OwnedDependencyChain {
-    vec: Vec<TaskId>,
-}
-
-impl OwnedDependencyChain {
-    #[inline]
-    #[must_use]
-    pub fn into_inner(self) -> Vec<TaskId> {
-        self.vec
-    }
-}
-
-impl std::fmt::Display for OwnedDependencyChain {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        for (i, task_id) in self.vec.iter().enumerate() {
-            if i > 0 {
-                write!(f, " -> ")?;
-            }
-            write!(f, "{task_id}")?;
-        }
-        Ok(())
     }
 }
